Refactor Storage and TaskList
